// apps/learn/core/entities/category/api/types.ts
import type { ProductDto } from '../../product/@x/category';

export type CategoryDto = {
  id: number;
  name: string;
  imageUrl: string[];
};

export type CategoryWithProductsDto = CategoryDto & {
  products: ProductDto[];
};

export type CategoryDetailsRequestArgs = {
  categoryId: number;
  sortBy?: string;
};